[refactor]convert function of Form to FormWorkData
